move feather/build.mjs -> scripts/feather.mjs
